pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cpdbench"
version = "0.0.4"
description = 'A benchmarking framework to run and evaluate changepoint detection algorithms.'
readme = "README.md"
requires-python = ">=3.9"
license = "BSD-2-Clause"
keywords = []
authors = [
  { name = "Dominik Sauerer", email = "dominik.sauerer@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "numpy",
  "pyyaml",
  "tqdm",
  "changepoynt==0.0.7"
]

[project.urls]
Documentation = "https://github.com/Lucew/CPD-Bench#readme"
Issues = "https://github.com/Lucew/CPD-Bench/issues"
Source = "https://github.com/Lucew/CPD-Bench"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/cpdbench tests}"

[tool.coverage.run]
source_pkgs = ["cpdbench", "tests"]
branch = true
parallel = true
omit = [
  "src/cpdbench/__about__.py",
]

[tool.coverage.paths]
cpdbench = ["src/cpdbench", "*/cpdbench/src/cpdbench"]
tests = ["tests", "*/cpdbench/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.hatch.build.targets.sdist]
exclude = [
  "dachboden.py",
  ".gitignore",
  "main.py",
  "config.yml"
]
